requirements.txt
bs4
selenium
langdetect
psycopg2-binary
pinecone-client
courlan